secure_file_sharing/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-sendgrid-v5==1.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyotp==2.9.0
python-http-client==3.3.7
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
